backend/requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
anthropic>=0.18.0
groq>=0.4.0
pydantic>=2.5.0
